Added Answer subclasses and separated submission and answer Dtos from their model objects.

+++ src/main/java/org/citizensense/model/CompassAnswer.java
@@ -3,23 +3,23 @@ package org.citizensense.model;
 public class CompassAnswer extends Answer {
 
 	
-	public Integer heading;
+	public Float heading;
 	
 	
 	public CompassAnswer(){
 		super();
 	}
 	
-	public CompassAnswer(Integer id, String answer_type, Integer q_id, Integer sub_id, Integer heading) {
+	public CompassAnswer(Integer id, String answer_type, Integer q_id, Integer sub_id, Float heading) {
 		super(id, answer_type, q_id, sub_id);
 		this.heading = heading;
 	}
 	
 	
-	public void setHeading(Integer h){
+	public void setHeading(float h){
 		this.heading = h;
 	}
-	public Integer getHeading(){
+	public Float getHeading(){
 		return heading;
 	}
